demo app updated
