+++ Demo/Demo/SceneDelegate.swift
@@ -34,12 +34,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 			)
 			
 			let apiStore = APIStore(uploadcare: uploadcare)
-			let uploader = Uploader(uploadcare: uploadcare)
 			
 		    window.rootViewController = UIHostingController(
 				rootView: contentView
 					.environmentObject(apiStore)
-					.environmentObject(uploader)
 			)
 		    self.window = window
 		    window.makeKeyAndVisible()
